+++ lib/feature/account/type/type.dart
@@ -1,3 +1,3 @@
 export 'email.dart';
-export 'loginId.dart';
+export 'login_id.dart';
 export 'password.dart';
